Stage 0: Tilemap with clickable images.
The tilemap currently has clickable images, and renders
in a grid.
The only type of interaction working is the "travel",
i.e. you can click on a Location to travel to it.

The next step will be stubbing the images with buttons.
Searching for images takes a long time, and I will need
to figure out things like how to scale the images so they
are all proportional to each other.
Having buttons with words will save a lot of time.